backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-genai>=1.65.0
Pillow>=12.0.0
python-dotenv>=1.2.1
python-multipart>=0.0.20
